refactor(segments): unify segment model and improve event details display
- Removed separate Firestore segment type; segments now use a single shared model
- Segments are stored directly on the event record instead of a separate subcollection
- Distance values updated to use higher precision throughout the create event flow
- Event details screen now shows both planned and actual segment data side by side
- Active events show goal time and distance per segment

+++ Modules/Dashboard/Home/NewRun/Views/DistanceStepView.swift
@@ -44,7 +44,7 @@ struct DistanceStepView: View {
 					set: { (newValue: Int) in
 						let fractional = viewModel.distance - floor(viewModel.distance)
 						let clampedInt = max(0, min(999, newValue))
-						viewModel.distance = Float(clampedInt) + fractional
+						viewModel.distance = Double(clampedInt) + fractional
 					}
 				)) {
 					ForEach(1...999, id: \.self) { intVal in
@@ -78,7 +78,7 @@ struct DistanceStepView: View {
 					set: { (newValue: Int) in
 						let clampedHundredths = max(0, min(99, newValue))
 						let intPart = Int(viewModel.distance)
-						viewModel.distance = Float(intPart) + Float(clampedHundredths) / 100.0
+						viewModel.distance = Double(intPart) + Double(clampedHundredths) / 100.0
 					}
 				)) {
 					ForEach(0...99, id: \.self) { frac in
